+++ bin/shiori/src/lib.rs
@@ -1,5 +1,7 @@
+mod app;
 pub mod commands;
 mod i18n;
 pub mod inspect;
 
+pub use app::ShioriApp;
 pub use shiori_plugin::async_trait;