+++ bin/shiori/src/main.rs
@@ -1,11 +1,38 @@
 use clap::Parser;
 use clap_handler::Handler;
 use shiori::commands::ShioriArgs;
-use tracing_subscriber::filter::LevelFilter;
+use std::fmt;
+use tracing::level_filters::LevelFilter;
+use tracing_subscriber::fmt::time::FormatTime;
+
+struct TimeOnly;
+
+impl FormatTime for TimeOnly {
+    fn format_time(&self, w: &mut tracing_subscriber::fmt::format::Writer<'_>) -> fmt::Result {
+        let now = std::time::SystemTime::now()
+            .duration_since(std::time::UNIX_EPOCH)
+            .unwrap();
+        let secs = now.as_secs();
+        let millis = now.subsec_millis();
+
+        let hours = (secs / 3600) % 24;
+        let minutes = (secs / 60) % 60;
+        let seconds = secs % 60;
+
+        write!(
+            w,
+            "{:02}:{:02}:{:02}.{:03}",
+            hours, minutes, seconds, millis
+        )
+    }
+}
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
     tracing_subscriber::fmt()
+        .with_timer(TimeOnly)
+        .with_target(false)
+        .with_level(true)
         .with_env_filter(
             tracing_subscriber::EnvFilter::builder()
                 .with_default_directive(LevelFilter::INFO.into())